Fixing some methods in QuizController and upgrading functionality of old ones.

+++ VoteAndQuizWebApi/Controllers/QuizzesController.cs
@@ -8,6 +8,9 @@ using VoteAndQuizWebApi.Dto;
 
 using VoteAndQuizWebApi.Models;
 using VoteAndQuizWebApi.Repository.IRepository;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using System.Security.Claims;
 
 namespace VoteAndQuizWebApi.Controllers
 {
@@ -18,12 +21,16 @@ namespace VoteAndQuizWebApi.Controllers
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
         private readonly IQuizRepository _quizRepository;
-        public QuizzesController(IUnitOfWork unitOfWork, IQuizRepository quizRepository, IMapper mapper)
+        private readonly UserManager<User> _userManager;
+        public QuizzesController(IUnitOfWork unitOfWork, IQuizRepository quizRepository, IMapper mapper
+            , UserManager<User> userManager)
         {
             _unitOfWork = unitOfWork;
             _quizRepository = quizRepository;
             _mapper = mapper;
+            _userManager = userManager;
         }
+
         [HttpGet]
         public IActionResult Index() //Lists all quizzes on the main quiz page
         {
@@ -31,36 +38,65 @@ namespace VoteAndQuizWebApi.Controllers
             if (!ModelState.IsValid) return BadRequest(ModelState);
             return Json(quizzes);
         }
-        [HttpGet("{id}")]
+
+        [HttpGet("Details/{id}")]
         [ProducesResponseType(200)]
         public IActionResult Details(int? id) //You should be able to access Details page to vote for a quiz option.
         {
             if (id == null)
                 return BadRequest();
 
-           
             //Details Method Dto ***
             var quiz = _mapper.Map<QuizForIndexMethodDTO>(_unitOfWork.Quiz.Get(q => q.Id == id, "Options"));
-            if (quiz.IsDeleted == true)
-                return BadRequest();
-            
+
             if (quiz == null)
                 return NotFound();
 
+            if (quiz.IsDeleted == true)
+                return BadRequest();
+
             return Json(quiz);
         }
-        [HttpPost]
-        public IActionResult
-            Create([FromBody] QuizForCreateMethodDTO quiz) //TODO: fix dto if needed
+
+        [HttpPost("Create")]
+        public IActionResult Create([FromBody] QuizForCreateMethodDTO quizDto)
         {
-            if (quiz == null)
+            // Validate input
+            if (quizDto == null || quizDto.Options == null || quizDto.Options.Count < 2)
+            {
+                ModelState.AddModelError("", "Quiz must have at least 2 options.");
                 return BadRequest(ModelState);
+            }
 
-            var newQuiz = _mapper.Map<Quiz>(quiz);
+            string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            User user = _unitOfWork.User.Get(u => u.Id == userId);
+            if (user == null) return Unauthorized("Log in to create a quiz");
 
-            if (!ModelState.IsValid)
+            // Check if QuizEndDate is less than one day from now
+            if (quizDto.QuizEndDate < DateTime.UtcNow.AddDays(1))
+            {
+                ModelState.AddModelError("", "Quiz end date must be at least one day from today.");
                 return BadRequest(ModelState);
-            var quizObj = _quizRepository.Get(q => q.Name.Trim().ToUpper() == quiz.Name.TrimEnd().ToUpper());
+            }
+
+            var newQuiz = new Quiz
+            {
+                Name = quizDto.Name,
+                UpdatedAt = DateTime.UtcNow.AddHours(3),
+                DeletedAt = null,
+                CreatedAt = DateTime.UtcNow.AddHours(3),
+                CreatorId = userId,
+                QuizEndDate = quizDto.QuizEndDate == DateTime.MinValue ? DateTime.UtcNow.AddDays(14) : quizDto.QuizEndDate,
+                quizVotes = quizDto.quizVotes,
+                Options = _mapper.Map<List<UserQuizAnswer>>(quizDto.Options),
+                CorrectOption = _mapper.Map<QuizOption>(quizDto.CorrectOption),
+                IsActive = true,
+                IsDeleted = false,
+                ShowQuiz = true,
+
+            };
+
+            var quizObj = _quizRepository.Get(q => q.Name.Trim().ToUpper() == quizDto.Name.TrimEnd().ToUpper());
             if (quizObj != null)
             {
                 ModelState.AddModelError("", "Quiz already exists");
@@ -73,9 +109,9 @@ namespace VoteAndQuizWebApi.Controllers
                 return StatusCode(500, ModelState);
             }
 
-
             return Ok("Successfully created");
         }
+
         [HttpDelete("Delete/{quizId}")]
         public IActionResult DeleteQuiz(int quizId)
         {
@@ -88,6 +124,17 @@ namespace VoteAndQuizWebApi.Controllers
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            User user = _unitOfWork.User.Get(u => u.Id == userId);
+            if (user == null) return Unauthorized();
+
+            // Check if the logged-in user is the creator of the quiz
+            if (quizToDelete.CreatorId != userId)
+            {
+                return Unauthorized("You are not authorized to delete this quiz.");
+            }
+
             if (!_quizRepository.DeleteQuiz(quizToDelete))
             {
                 ModelState.AddModelError("", "Something went wrong deleting quiz");
@@ -95,7 +142,7 @@ namespace VoteAndQuizWebApi.Controllers
 
             return Ok("Successfully deleted");
         }
-        [HttpPut("{quizId}")]
+        [HttpPut("{quizId}")] //this does not make sense... add voting endpoint
         public IActionResult UpdateQuiz(int quizId)//updating quiz happens when someone votes for a quiz option
         {
             var quiz = _unitOfWork.Quiz.Get(q => q.Id == quizId);
@@ -113,15 +160,26 @@ namespace VoteAndQuizWebApi.Controllers
         [HttpPost("Finish/{quizId}")]
         public IActionResult FinishQuiz(int quizId) //only creator can finish it
         {
+            string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            User user = _unitOfWork.User.Get(u => u.Id == userId);
+            if (user == null) return Unauthorized("Log in to finish this quiz");
+
             var quiz = _unitOfWork.Quiz.Get(q => q.Id == quizId);
             if (quiz == null)
                 return BadRequest(ModelState);
+
+            if (quiz.CreatorId != userId)
+            {
+                return Unauthorized("You are not authorized to finish this quiz.");
+            }
+
             quiz.QuizEndDate = DateTime.UtcNow.AddHours(3);
             quiz.IsActive = false;
             quiz.ShowQuiz = false;
             quiz.IsDeleted = true;
             quiz.UpdatedAt = DateTime.UtcNow.AddHours(3);
             quiz.DeletedAt = DateTime.UtcNow.AddHours(3);
+
             _unitOfWork.Quiz.Update(quiz);
             _unitOfWork.Save();
